--- api/helpers/postsHelper.js
@@ -1,16 +1,0 @@
-const db = require('../../dbConfig');
-
-// USER HELPERS
-// ==============================================
-module.exports = {
-  get: function(id) {
-    let query = db('posts');
-    if (id) query.where('id', Number(id)).first();
-    return query;
-  },
-  insert: function(post) {
-    return db('posts')
-      .insert(post)
-      .then(ids => ({ id: ids[0] }));
-  }
-};
