+++ components/auth/index.js
@@ -1,0 +1,5 @@
+const authRoutes = require('./authRoutes');
+
+module.exports = (server) => {
+  server.use('/auth', authRoutes)
+}
