+++ components/posts/index.js
@@ -1,0 +1,5 @@
+const postRoutes = require('./postRoutes');
+
+module.exports = (server) => {
+  server.use('/api/posts', postRoutes)
+}
