+++ components/services/index.js
@@ -1,0 +1,5 @@
+const servicesRoutes = require('./servicesRoutes');
+
+module.exports = (server) => {
+  server.use('/api', servicesRoutes);
+}
