+++ index.js
@@ -2,7 +2,8 @@ require('dotenv').config();
 const server = require('express')();
 require('./config/passport');
 require('./middleware/index')(server);
-require('./api/Routes')(server);
+require('./components')(server);
+
 const port = process.env.PORT || 8000;
 server.get('/', (req, res) => {
   res.send('localhost up & alive');
